clean up merkle tree
